Add "I have" capability; Add files and url as db attribute; init()
Allow user to mark albums, recordings, publications, and dances as "I
Have", and let them tag these things with a custom, user-defined string.

Give the db the dbFile, saveFile, and dbURL as attributes so we don't
have to create those in every function that wants to use them.

Create an init function to initiate a connection with the database. Also
use this funciton after updating the database to open a new connection
with what was updated.